Move SystemTime to common module
